1_test0_3

+++ modules/task_1/Ivlev_A_min_val_in_rows/main.cpp
@@ -1,9 +1,8 @@
-// Copyright 2022 Ivlev A
+  // Copyright 2022 Ivlev A
 #include <gtest/gtest.h>
 #include <vector>
 #include "./min_val_in_rows.h"
 #include <gtest-mpi-listener.hpp>
-//#include <iostream>
 
 
 TEST(test_MPI, test_min_0) {
@@ -23,7 +22,7 @@ TEST(test_MPI, test_min_0) {
     if(rank == 0){   
         int* global_min_local = getMatrixMinbyRow(global_matrix, m, n);
 
-        for(size_t i = 0; i < m; i++) {
+        for(int i = 0; i < m; i++) {
             ASSERT_EQ(global_min_local[i], global_min[i]);
         }
         
@@ -50,7 +49,7 @@ TEST(test_MPI, test_min_1) {
     if(rank == 0){   
         int* global_min_local = getMatrixMinbyRow(global_matrix, m, n);
 
-        for(size_t i = 0; i < m; i++) {
+        for(int i = 0; i < m; i++) {
             ASSERT_EQ(global_min_local[i], global_min[i]);
         }
         
@@ -77,7 +76,7 @@ TEST(test_MPI, test_min_2) {
     if(rank == 0){   
         int* global_min_local = getMatrixMinbyRow(global_matrix, m, n);
 
-        for(size_t i = 0; i < m; i++) {
+        for(int i = 0; i < m; i++) {
             ASSERT_EQ(global_min_local[i], global_min[i]);
         }
         
@@ -104,7 +103,7 @@ TEST(test_MPI, test_min_3) {
     if(rank == 0){   
         int* global_min_local = getMatrixMinbyRow(global_matrix, m, n);
 
-        for(size_t i = 0; i < m; i++) {
+        for(int i = 0; i < m; i++) {
             ASSERT_EQ(global_min_local[i], global_min[i]);
         }
         
@@ -131,7 +130,7 @@ TEST(test_MPI, test_min_4) {
     if(rank == 0){   
         int* global_min_local = getMatrixMinbyRow(global_matrix, m, n);
 
-        for(size_t i = 0; i < m; i++) {
+        for(int i = 0; i < m; i++) {
             ASSERT_EQ(global_min_local[i], global_min[i]);
         }
